Вряд ли это часть сервиса для работы с пользователями. Скорее это часть сервиса отвечающего за тестирование
